refactor(app): extract CollectionRoutes from App.tsx (#316 phase 6)
Twenty screens on the collection/shop/profile side: player, collection-tabs,
collection, augments, the four ShopScreen variants, deckbuilder, pack,
inventory, achievements, hall-of-achievements, the two home-shelf tabs,
heroCards, codex, playerstats, chronicle and endlessleaderboard.

AppContextValue gains setCommander, packs, shopBuildingId, shopTappedNpc and
the three shop/pack handlers.

App.tsx: 3,370 -> 3,224 lines. Build green, chunk boundaries unchanged,
2107 tests passing.
